feat: add user permissions

+++ estoque/blueprints/restapi/resouces/ProductType/ProductTypeResource.py
@@ -2,7 +2,7 @@ from flask import jsonify, request
 from estoque.blueprints.restapi.httpMessages.httpSucess import httpSuccess
 from estoque.blueprints.restapi.httpMessages.httpError import httpError
 from estoque.blueprints.restapi.requests.requestChecker import requestChecker
-from estoque.models import ProductTypes
+from estoque.models import ProductTypes, UserPermission
 from flask_restful import Resource
 from ...httpMessages.httpError import httpError
 from estoque.ext.database import db
@@ -29,18 +29,40 @@ class ProductTypeItemResource(Resource):
         if productstypes:
             return jsonify(productstypes.to_dict())
         else:
-            return httpError("Product type does not exist", 404)
+            return httpError("Product type does not exist", 403)
 
-    def delete(self, producttype_id):
+    def delete(self, producttype_id, user_id):
+        user_permission = UserPermission.query.filter_by(
+            user_id=user_id).first()
+
+        if not user_permission:
+            return httpError("You have no permissions", 403)
+
+        if user_permission.permission not in ['delete_user_type', 'administrador']:
+            return httpError("You don't have permission to delete a product type", 403)
+
         productstypes = ProductTypes.query.filter_by(ID=producttype_id).first()
+
         if not productstypes:
             return httpError('Product type does not exist', 404)
+
         db.session.delete(productstypes)
+
         db.session.commit()
+
         return httpSuccess('Product type deleted successfully')
 
-    def put(self, producttype_id):
+    def put(self, producttype_id, user_id):
         productstypes = ProductTypes.query.filter_by(ID=producttype_id).first()
+
+        user_permission = UserPermission.query.filter_by(
+            user_id=user_id).first()
+
+        if not user_permission:
+            return httpError("You have no permissions", 403)
+
+        if user_permission.permission not in ['update_user_type', 'administrador']:
+            return httpError("You don't have permission to update a product type", 403)
 
         if not productstypes:
             return httpError('Product type does not exist', 404)
